MEXSVR-410 将来始まる軽減税率を見据えて、各種テーブルの税率のカラム(TaxRate)のデータ型を揃える
git-svn-id: http://svnsin.ncr.com/svn/repos/pa/res/RES%20MeX/Base/resTransaction/branches/3.3.2@76872 9234da4d-b7fe-4b44-acf3-10f21a033f33

+++ src/ncr/res/mobilepos/pricing/dao/SQLServerItemDAO.java
@@ -205,7 +205,7 @@ public class SQLServerItemDAO extends AbstractDao implements IItemDAO {
                 } else {
                 	searchedItem.setTaxType(result.getInt(result.findColumn("dptTaxType")));
                 }
-                searchedItem.setTaxRate(result.getDouble(result.findColumn("TaxRate")));
+                searchedItem.setTaxRate(result.getInt(result.findColumn("TaxRate")));
                 searchedItem.setSubNum1(result.getInt(result.findColumn("SubNum1")));
                 searchedItem.setSubNum2(result.getInt(result.findColumn("SubNum2")));
 
@@ -845,7 +845,7 @@ public class SQLServerItemDAO extends AbstractDao implements IItemDAO {
                 searchedItem.setDescription(description);
 
                 searchedItem.setTaxType(result.getInt(result.findColumn("TaxType")));
-                searchedItem.setTaxRate(result.getDouble(result.findColumn("TaxRate")));
+                searchedItem.setTaxRate(result.getInt(result.findColumn("TaxRate")));
                 searchedItem.setSubNum1(result.getInt(result.findColumn("SubNum1")));
                 searchedItem.setSubNum2(result.getInt(result.findColumn("SubNum2")));
